Verkort ongedaan-maken snackbar naar SnackbarDuration.Short (~4s)
Long (~10s) was langer dan nodig; Short is de standaard, ingebouwde
Material3-duur en voldoet prima i.p.v. een handmatige timeout te
bouwen voor een exacte tijdsduur.

+++ app/src/main/java/com/dtraas/boodschapbeheer/ui/shoppinglist/ShoppingListScreen.kt
@@ -171,13 +171,13 @@ fun ShoppingListScreen() {
             fun deleteWithUndo(item: ShoppingListItemEntity) {
                 viewModel.removeItem(item.id)
                 coroutineScope.launch {
+                    // showSnackbar defaults to SnackbarDuration.Indefinite whenever an
+                    // actionLabel is set, so without an explicit duration the "ongedaan
+                    // maken" snackbar would never auto-dismiss.
                     val result = snackbarHostState.showSnackbar(
                         message = removedFormat.format(item.name),
                         actionLabel = undoLabel,
-                        // showSnackbar defaults to SnackbarDuration.Indefinite whenever an
-                        // actionLabel is set, so without this the "ongedaan maken" snackbar
-                        // would never auto-dismiss.
-                        duration = SnackbarDuration.Long,
+                        duration = SnackbarDuration.Short,
                     )
                     if (result == SnackbarResult.ActionPerformed) {
                         viewModel.restoreItem(item)
